requirements.txt
farm-ng-amiga
fastapi
uvicorn[standard]
opencv-python
numpy
ultralytics
